+++ src/backend/efiling-reviewer-api/src/main/java/ca/bc/gov/open/jag/efilingreviewerapi/extract/DocumentsApiDelegateImpl.java
@@ -1,5 +1,6 @@
 package ca.bc.gov.open.jag.efilingreviewerapi.extract;
 
+import ca.bc.gov.open.efilingdiligenclient.diligen.DiligenService;
 import ca.bc.gov.open.jag.efilingreviewerapi.api.DocumentsApiDelegate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
