+++ src/backend/libs/efiling-diligen-client/src/main/java/ca/bc/gov/open/efilingdiligenclient/diligen/DiligenService.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.open.jag.efilingreviewerapi.extract;
+package ca.bc.gov.open.efilingdiligenclient.diligen;
 
 import org.springframework.web.multipart.MultipartFile;
